leetcode/02-Add-Two-Numbers/go/main_test.go
package main

import (
	"reflect"
	"testing"
)

func TestMain(t *testing.T) {
	tests := []struct {
		name string
		l1   []int
		l2   []int
		want []int
	}{
		{
			"case 1",
			[]int{2, 4, 3},
			[]int{5, 6, 4},
			[]int{7, 0, 8},
		},
		{
			"case 2",
			[]int{0},
			[]int{0},
			[]int{0},
		},
		{
			"case 3",
			[]int{9, 9, 9, 9, 9, 9, 9},
			[]int{9, 9, 9, 9},
			[]int{8, 9, 9, 9, 0, 0, 0, 1},
		},
	}

	for _, tt := range tests {
		tt := tt
		t.Run(tt.name, func(t *testing.T) {
			got := addTwoNumbers(convertToListNodeFromList(tt.l1), convertToListNodeFromList(tt.l2))
			if !reflect.DeepEqual(convertToListFromNodeList(got), tt.want) {
				t.Errorf("name %v, got %v, want %v", tt.name, got, tt.want)
			}
		})
	}
}
